% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivativeHoldingsPullAndScrape.R
\name{derivativeHoldingsPullAndScrape}
\alias{derivativeHoldingsPullAndScrape}
\title{derivativeTransactionsPullAndScrape}
\usage{
derivativeHoldingsPullAndScrape(
  quarter,
  year,
  form,
  name,
  email,
  footnoteKeywords = NA,
  issuerKeywords = NA,
  issuerTradingSymbol = NA,
  rptOwnerKeywords = NA
)
}
\arguments{
\item{quarter}{Specify the financial reporting quarter. Options are 1, 2, 3,
and 4.}

\item{year}{Specify the calendar year. Options range from 2004 to present.}

\item{form}{Specify the form type. Options are 4 or 5.}

\item{name}{Specify your name. This is required by the SEC.}

\item{email}{Specify your email. This is required by the SEC.}

\item{footnoteKeywords}{The character vector containing any specified key
words to be searched in the form's footnotes. Default is NA.}

\item{issuerKeywords}{The character vector containing any specified key
words to be searched within the issuer block. Criteria can include the firms
CIK or name. Values must be capitalized to match against SEC values.
Default value is NA.}

\item{issuerTradingSymbol}{The character vector containing any specified
stock tickers.  Values must be capitalized to match against SEC values.
Default value is NA.}

\item{rptOwnerKeywords}{The character vector contianing any specified key
words to be searched within the rptOwner block.  Values must be capitalized
to match against SEC values. Additionally, the format is LAST NAME, FIRST
NAME. The criteria can include the individuals CIK, address, or name.
Default value is NA.}
}
\value{
A dataframe containing containing the scraped information where
each row represents a derivative holding The holding
observations can be grouped by form through the URL variable within the
dataframe.
}
\description{
The function takes the specifieid quarter (1-4), year (2004-
2021), and form type (4 or 5) and returns a dataframe of all
the derivative holdings that meet the keyword crtiera from the
specified financial quarter. If no criteria is set, all holdings from
the financial quarter are scraped. Note that the logic for the
keywords is "OR." Thus if a user sets multiple footnoteKeywords and a
keyword for issuerKeywords, if a holding has only one of the
footnoteKeywords and none of the issuerKeywords, the holding will
still be scraped.
}
