% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivativeHoldingsScrape.R
\name{derivativeHoldingsScrape}
\alias{derivativeHoldingsScrape}
\title{derivativeHoldingsScrape}
\usage{
derivativeHoldingsScrape(
  index,
  form,
  name,
  email,
  footnoteKeywords = NA,
  issuerKeywords = NA,
  issuerTradingSymbol = NA,
  rptOwnerKeywords = NA
)
}
\arguments{
\item{index}{A local object that was returned by utilizing the
\code{\link{secUrlDownload}}. It contains the URLs necessary to grab each
text version of the filed Form 4 or 5.}

\item{form}{Define if the form is a Form 4 or Form 5. Must match the type
used in \code{\link{secUrlDownload}}.}

\item{name}{Specify your name. This is required by the SEC.}

\item{email}{Specify your email. This is required by the SEC.}

\item{footnoteKeywords}{The character vector containing any specified key
words to be searched in the form's footnotes. Default is NA.}

\item{issuerKeywords}{The character vector containing any specified key
words to be searched within the issuer block. Criteria can include the firms
CIK or name. Values must be capitalized to match against SEC values.
Default value is NA.}

\item{issuerTradingSymbol}{The character vector containing any specified
stock tickers.  Values must be capitalized to match against SEC values.
Default value is NA.}

\item{rptOwnerKeywords}{The character vector containing any specified key
words to be searched within the rptOwner block.  Values must be capitalized
to match against SEC values. Additionally, the format is LAST NAME, FIRST
NAME. The criteria can include the individuals CIK, address, or name.
Default value is NA.}
}
\value{
A dataframe containing containing the scraped information where
each row represents a derivative holding The holding observations
can be grouped by form through the URL variable within the dataframe.
}
\description{
The function takes the dataframe returned by
\code{\link{secUrlDownload}} and scrapes and compiles a dataframe of all
the derivative holdings that meet the keyword criteria. If no criteria
is set, all holdings from the dataframe returned by
\code{\link{secUrlDownload}} are scraped. Note that the logic for the
keywords is "OR." Thus if a user sets multiple footnoteKeywords and a
keyword for issuerKeywords, if a holding has only one of the
footnoteKeywords and none of the issuerKeywords, the holding will
still be scraped.
}
\examples{
\donttest{
dat <- derivativeHoldingsScrape(index = tempIndex, form = 4,
name = "Your Name", email = "yourEmail@yourEmail.com",
issuerKeywords = c("AMAZON", "MICROSOFT"))
}
}
